Refactor: Change annotation deprecated 'where' to 'SQLRestriction' and Trim white spaces #26
