+++ src/main/java/org/certis/studyplatform/auth/infrastructure/persistence/AuthEntity.java
@@ -10,13 +10,13 @@ import org.certis.studyplatform.member.infrastructure.persistence.entity.MemberE
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 import org.hibernate.annotations.SQLDelete;
-import org.hibernate.annotations.Where;
+import org.hibernate.annotations.SQLRestriction;
 
 import java.time.OffsetDateTime;
 
 /**
  * Auth Entity
- * 
+ *
  * 회원 인증 정보를 저장하는 JPA Entity
  * Member와 1:1 관계를 가짐
  */
@@ -25,16 +25,16 @@ import java.time.OffsetDateTime;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
 @SQLDelete(sql = "UPDATE auth SET deleted_at = NOW() WHERE member_id = ?")
-@Where(clause = "deleted_at IS NULL")
+@SQLRestriction("deleted_at IS NULL")
 public class AuthEntity {
 
     @Id
     @Column(name = "member_id", nullable = false, unique = true)
     private Long memberId;  // member_id를 PK로 사용
-    
+
     @Column(name = "account_number", nullable = false, unique = true)
     private String accountNumber;
-    
+
     // Member와의 관계 매핑 추가
     @OneToOne
     @JoinColumn(name = "member_id", insertable = false, updatable = false)
@@ -55,7 +55,7 @@ public class AuthEntity {
     private OffsetDateTime deletedAt;
 
     @Builder(toBuilder = true)
-    private AuthEntity(String accountNumber, String password, OffsetDateTime createdAt, 
+    private AuthEntity(String accountNumber, String password, OffsetDateTime createdAt,
                       OffsetDateTime updatedAt, OffsetDateTime deletedAt, Long memberId) {
         this.accountNumber = accountNumber;
         this.password = password;
@@ -71,4 +71,4 @@ public class AuthEntity {
         }
         this.password = newPassword;
     }
-} 
+}
