+++ src/main/java/org/certis/studyplatform/member/infrastructure/persistence/entity/MemberEntity.java
@@ -115,4 +115,4 @@ public class MemberEntity {
                 ", createdAt=" + createdAt +
                 '}';
     }
-} 
+}
