+++ src/main/java/org/certis/studyplatform/member/infrastructure/persistence/entity/MemberPenaltyEntity.java
@@ -11,7 +11,7 @@ import java.time.OffsetDateTime;
 
 /**
  * Member Penalty Entity
- * 
+ *
  * 회원 패널티 정보를 저장하는 JPA Entity
  * Member와 1:1 관계
  */
@@ -47,4 +47,4 @@ public class MemberPenaltyEntity {
         this.updatedAt = updatedAt;
         this.reason = reason;
     }
-} 
+}
