+++ src/main/java/org/certis/studyplatform/study/infrastructure/persistence/entity/StudyAttachedEntity.java
@@ -1,4 +1,4 @@
-package org.certis.studyplatform.study.infrastructure.persistence;
+package org.certis.studyplatform.study.infrastructure.persistence.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -8,22 +8,22 @@ import lombok.NoArgsConstructor;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 import org.hibernate.annotations.SQLDelete;
-import org.hibernate.annotations.Where;
+import org.hibernate.annotations.SQLRestriction;
 
 import java.time.OffsetDateTime;
 
 /**
- * Study Participant Entity
- * 
- * 스터디 참가자를 저장하는 JPA Entity
+ * study Attached Entity
+ *
+ * 프로젝트 첨부파일을 저장하는 JPA Entity
  */
 @Entity
-@Table(name = "study_participant")
+@Table(name = "study_attached")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-@SQLDelete(sql = "UPDATE study_participant SET deleted_at = NOW() WHERE id = ?")
-@Where(clause = "deleted_at IS NULL")
-public class StudyParticipantEntity {
+@SQLDelete(sql = "UPDATE study_attached SET deleted_at = NOW() WHERE id = ?")
+@SQLRestriction("deleted_at IS NULL")
+public class StudyAttachedEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -34,6 +34,18 @@ public class StudyParticipantEntity {
 
     @Column(name = "member_id", nullable = false)
     private Long memberId;
+
+    @Column(name = "attached_url", nullable = false)
+    private String attachedUrl;
+
+    @Column(name = "name", nullable = false)
+    private String name;
+
+    @Column(name = "type", nullable = false)
+    private String type;
+
+    @Column(name = "size", nullable = false)
+    private String size;
 
     @CreationTimestamp
     @Column(name = "created_at", nullable = false, updatable = false)
@@ -47,13 +59,18 @@ public class StudyParticipantEntity {
     private OffsetDateTime deletedAt;
 
     @Builder(toBuilder = true)
-    private StudyParticipantEntity(Long id, Long studyId, Long memberId,
-                                  OffsetDateTime createdAt, OffsetDateTime updatedAt, OffsetDateTime deletedAt) {
+    private StudyAttachedEntity(Long id, Long studyId, Long memberId, String attachedUrl,
+                                String name, String type, String size,
+                                OffsetDateTime createdAt, OffsetDateTime updatedAt, OffsetDateTime deletedAt) {
         this.id = id;
         this.studyId = studyId;
         this.memberId = memberId;
+        this.attachedUrl = attachedUrl;
+        this.name = name;
+        this.type = type;
+        this.size = size;
         this.createdAt = createdAt;
         this.updatedAt = updatedAt;
         this.deletedAt = deletedAt;
     }
-} 
+}
